Moved some more deckchairs around
