+++ include/skyr/url.hpp
@@ -17,7 +17,7 @@
 #include <skyr/config.hpp>
 #include <skyr/version.hpp>
 #include <skyr/core/url_record.hpp>
-#include <skyr/core/url_error.hpp>
+#include <skyr/core/errors.hpp>
 #include <skyr/url_search_parameters.hpp>
 #include <skyr/network/ipv4_address.hpp>
 #include <skyr/network/ipv6_address.hpp>
