+++ src/url/url.cpp
@@ -6,11 +6,11 @@
 #include <cassert>
 #include <functional>
 #include <skyr/url.hpp>
-#include <skyr/core/url_parse.hpp>
-#include <skyr/core/url_serialize.hpp>
+#include <skyr/core/parse.hpp>
+#include <skyr/core/serialize.hpp>
 #include <skyr/percent_encoding/percent_encode_range.hpp>
 #include "core/url_parse_impl.hpp"
-#include "skyr/core/url_schemes.hpp"
+#include "skyr/core/schemes.hpp"
 
 namespace skyr {
 inline namespace v1 {
